Changes to default types to enable Tpetra (with int as local ordinal type)
Still need to use long long int for file size and buffer size in cine
reader because the files are too big

+++ src/core/DICe.h
@@ -85,7 +85,7 @@ namespace DICe{
   typedef float scalar_t;
 #endif
 
-typedef long int int_t;
+typedef int int_t;
 
 /// parameters (all lower case)
